requirements.txt
pyinstaller == 3.6
discord.py == 1.4.0
PyYAML == 5.3.1